requirements-security.txt
# 小红书 AI 运营系统 - 核心依赖
# Python 3.9+

# MCP 协议
mcp>=0.9.0

# 数据验证
pydantic>=2.0.0

# 定时任务
apscheduler>=3.10.0

# 安全相关
keyring>=20.0.0           # 系统密钥环管理
cryptography>=41.0.0      # 加密解密

# 异步文件 I/O
aiofiles>=23.0.0

# 数据库
aiosqlite>=0.19.0

# 缓存
redis>=5.0.0

# HTTP 客户端
httpx>=0.25.0

# 重试机制
tenacity>=8.2.0

# 熔断器
circuitbreaker>=2.0.0

# YAML 解析
pyyaml>=6.0

# 日志
structlog>=23.0.0

# 监控
prometheus-client>=0.19.0
